requirements.txt
openai
datetime
fastapi
starlette
requests
fastapi_sessions
app
pydantic
langchain
tiktoken 
pymupdf 
poppler-utils
pypdf
cryptocode
nltk
rich
uvicorn
pandas
numpy
nltk
msal==1.22.0
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.16.0
pandasai
mysql-connector-python